% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_element}
\alias{tidy_element}
\title{tidy_element}
\usage{
tidy_element(elems, depth = 0)
}
\arguments{
\item{elems}{a list of selenium-elements}

\item{depth}{a numeric value indicating how deep the search should go.}
}
\value{
A tibble with the elements attribute and a column for each layer containing the children from the corresponding layer.
}
\description{
tidy_element
}
