% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyweb-package.R
\docType{package}
\name{tidyweb-package}
\alias{tidyweb}
\alias{tidyweb-package}
\title{tidyweb: Easily Install and Load Modern Web-Scraping Packages}
\description{
Scraping the modern web requires having many tools in one's toolbox
and the 'tidyverse' only enables easy installation and loading of a few, core
ones. `tidyweb` bundles a more complete list of scraping tools to make it easier
to both install and use them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/tidyweb}
  \item Report bugs at \url{https://github.com/hrbrmstr/tidyweb/issues}
}

}
\keyword{internal}
