% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidywebverse_conflicts}
\alias{tidywebverse_conflicts}
\title{Conflicts between the tidywebverse and other packages}
\usage{
tidywebverse_conflicts()
}
\description{
This function lists all the conflicts between packages in the tidywebverse
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
tidywebverse_conflicts()
}
