% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidywebverse_deps}
\alias{tidywebverse_deps}
\title{List all tidywebverse dependencies}
\usage{
tidywebverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidywebverse packages.}
}
\description{
List all tidywebverse dependencies
}
