% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidywebverse_update}
\alias{tidywebverse_update}
\title{Update tidywebverse packages}
\usage{
tidywebverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
tidywebverse packages.}
}
\description{
This will check to see if all tidywebverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
tidywebverse_update()
}
}
